Bug fixes, updates

+++ discount/mkdio.h
@@ -5,7 +5,7 @@
 
 typedef void MMIOT;
 
-typedef unsigned long mkd_flag_t;
+typedef unsigned int mkd_flag_t;
 
 /* line builder for markdown()
  */
